backend/pyproject.toml
[tool.poetry]
name = "brainboost"
version = "0.1.0"
description = ""
authors = ["ArthurWilmer <arthurwrx@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
youtube-transcript-api = "^1.0.3"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.39"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.4.0"
python-multipart = "^0.0.20"
python-decouple = "^3.8"
nanoid = "^2.0.0"
openai = "0.28"
utils = "^1.0.2"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
